backend/requirements.txt
flask==2.0.1
gunicorn==20.1.0
pdfplumber==0.11.4
python-docx==1.1.2
pandas==2.2.2
scikit-learn==1.5.1
spacy==3.7.6
werkzeug==2.0.3
flask-cors==4.0.1
sentence-transformers==3.0.1
